requirements.txt
torch
transformers
scikit-learn
pandas
numpy
fastapi
uvicorn
flask
requests
pydantic
psutil
accelerate
newspaper3k
beautifulsoup4